requirements.txt
google-genai
pydantic
Pillow
fastapi
uvicorn
python-multipart
